+++ src/main/java/de/marhali/easyi18n/settings/presets/DefaultPreset.java
@@ -76,4 +76,9 @@ public class DefaultPreset implements ProjectSettings {
     public boolean isAssistance() {
         return true;
     }
+
+    @Override
+    public boolean isAlwaysFold() {
+        return false;
+    }
 }